requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
statsmodels>=0.14.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly>=5.15.0
kaleido==0.2.1

# Streamlit & Web App
streamlit>=1.28.0

# Export & Reporting
reportlab>=4.0.0
jinja2>=3.1.0

# File Processing
openpyxl>=3.1.0
xlrd>=2.0.0

# AI Integration
google-generativeai>=0.3.0

# Performance & Utilities
psutil>=5.9.0
python-dotenv>=1.0.0

# Optional: For better font support with reportlab on Linux
# (uncomment if you encounter font issues)
# pillow>=9.0.0

# VSCode/Pylance sẽ tự nhận diện nếu đúng môi trường venv.
